nonprofit_finance_db/receipt_scanning_tools/receipt_scanner/requirements.txt
# Receipt Scanner Standalone - Python Dependencies

# Core dependencies
google-generativeai>=0.3.0    # Gemini AI SDK
pydantic>=2.0.0                # Data validation
pillow>=10.0.0                 # Image processing
python-dotenv>=1.0.0           # Environment variables

# FastAPI and async
fastapi>=0.104.0              # Web framework
uvicorn>=0.24.0               # ASGI server
aiofiles>=23.2.0              # Async file operations

# Database
pymysql>=1.1.0                # MySQL connector
mysql-connector-python>=8.0.0 # Alternative MySQL connector

# Testing
pytest>=7.4.0                 # Test framework
pytest-asyncio>=0.21.0        # Async test support
httpx>=0.25.0                 # HTTP client for testing

# Optional but recommended
python-multipart>=0.0.6       # File upload support
sse-starlette>=1.6.5          # Server-sent events
